requirements.txt
streamlit==1.22.0
langchain==0.0.153
openai==0.27.5
anthropic==0.2.7
python-dotenv==1.0.0
